Fix buffer overflow bugs
In fastPathDepthConv (which combines sourceWord colour-depth conversion with
another fast path for another combinationRule at a constant colour depth)
and fastPathRightToLeft, it could overflow the temporary buffer and thereby
corrupt other local variables if the last chunk of a pixel row was 2048
bytes (or just under). This was most likely to happen with 32bpp destination
images and widths of about 512 pixels.