Detect and add a new fast path flag for effective-1bpp colour maps
Sometimes, colour maps are used such that all entries except the first
contain the same value. Combined with the fact that only source colour 0
uses colour map entry 0 (any other colours for which all non-0 bits would
otherwise be discarded during index generation are forced to use entry 1
instead), this effectively acts as a 2-entry (or 1bpp) map, depending on
whether the source colour is 0 or not. This is far more efficiently coded
in any fast path by a test against zero, than by a table lookup - it frees
up 2 KB, 16 KB or 128 KB of data cache space, depending on whether a 9-,
12- or 15-bit colour map was used. There is an up-front cost to scanning
the colour map to see if its entries are of this nature, however in most
"normal" colour maps, this scan will rapidly be aborted.